Refactor eval, add syntax transformer support
The evaluator function eval has been rewritten as nested cases. If we
have a form like (ident arg1 arg2 ...), we look up ~ident in the
environment and if it exists, we apply it to the form itself.

If that fails, we look ident up in the environment and, if it exists,
apply it as a function to the arguments in sequence.

Finally, if the function does not exist in the environment, we check
the builtins, and throw a runtime exception if this fails.